requirements.txt
# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# NLP Libraries
nltk>=3.8.0
spacy>=3.6.0
textblob>=0.17.0

# Machine Learning
xgboost>=2.0.0
lightgbm>=4.0.0
imbalanced-learn>=0.11.0

# Explainability
lime>=0.2.0
shap>=0.42.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
wordcloud>=1.9.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.3.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
